refactor(workflow): Move WorkflowAppGenerateTaskPipeline and WorkflowCycleManage to `core/workflow`
