+++ api/core/workflow/workflow_app_generate_task_pipeline.py
@@ -52,10 +52,10 @@ from core.app.entities.task_entities import (
     WorkflowTaskState,
 )
 from core.app.task_pipeline.based_generate_task_pipeline import BasedGenerateTaskPipeline
-from core.app.task_pipeline.workflow_cycle_manage import WorkflowCycleManage
 from core.ops.ops_trace_manager import TraceQueueManager
 from core.workflow.enums import SystemVariableKey
 from core.workflow.repository.workflow_node_execution_repository import WorkflowNodeExecutionRepository
+from core.workflow.workflow_cycle_manage import WorkflowCycleManage
 from extensions.ext_database import db
 from models.account import Account
 from models.enums import CreatedByRole
